requirements_venv.txt
pandas==2.2.3
matplotlib==3.9.4
psycopg2_binary==2.9.10
seaborn==0.13.2
sqlalchemy==2.0.39